refactor(errors): relocate _warn_symmetry_loss next to SymmetryLossWarning
