+++ src/flopscope/_symmetric.py
@@ -247,43 +247,7 @@ def is_symmetric(
 # Symmetry-loss warning helper
 # ---------------------------------------------------------------------------
 
-
-_FLOPSCOPE_PKG_DIR = os.path.dirname(os.path.abspath(__file__))
-
-
-def _user_stacklevel() -> int:
-    """Stacklevel for :func:`warnings.warn` that points to the first frame
-    outside the ``flopscope`` package — i.e. the user's call site.
-
-    Walks the active call stack starting from the caller of
-    :func:`_warn_symmetry_loss`. Robust to changes in the number of
-    decorator/wrapper layers between user code and the warn site.
-    """
-    frame = sys._getframe(2)
-    level = 2
-    while frame is not None:
-        if not frame.f_code.co_filename.startswith(_FLOPSCOPE_PKG_DIR):
-            return level
-        frame = frame.f_back
-        level += 1
-    return 3
-
-
-def _warn_symmetry_loss(
-    lost_dims: list[tuple[int, ...]],
-    reason: str,
-) -> None:
-    """Emit a :class:`SymmetryLossWarning` if warnings are enabled."""
-    if not get_setting("symmetry_warnings"):
-        return
-    dim_str = ", ".join(str(g) for g in lost_dims)
-    warnings.warn(
-        f"Symmetry lost along dims {dim_str}: {reason}. "
-        "Use as_symmetric() to re-tag if you know the result is symmetric. "
-        "Suppress with flops.configure(symmetry_warnings=False).",
-        SymmetryLossWarning,
-        stacklevel=_user_stacklevel(),
-    )
+from flopscope.errors import _warn_symmetry_loss  # re-exported for back-compat
 
 
 # ---------------------------------------------------------------------------
